+++ GameLib.Plugin/GameLib.Plugin.Gog/GogGameFactory.cs
@@ -26,6 +26,7 @@ internal static class GogGameFactory
             .Select(gameId => LoadFromRegistry(launcher, gameId))
             .Where(game => game is not null)
             .Select(game => AddLauncherId(launcher, game!))
+            .Select(game => AddExecutables(launcher, game!))
             .ToList()!;
     }
 
@@ -35,6 +36,22 @@ internal static class GogGameFactory
     private static GogGame AddLauncherId(ILauncher launcher, GogGame game)
     {
         game.LauncherId = launcher.Id;
+        return game;
+    }
+
+    /// <summary>
+    /// Find executables within the install directory
+    /// </summary>
+    private static GogGame AddExecutables(ILauncher launcher, GogGame game)
+    {
+        if (launcher.LauncherOptions.SearchExecutables)
+        {
+            var executables = PathUtil.GetExecutables(game.InstallDir);
+
+            executables.AddRange(game.Executables);
+            game.Executables = executables.Distinct(StringComparer.OrdinalIgnoreCase).ToList();
+        }
+
         return game;
     }